+++ Sources/TuistCache/ContentHashing/CacheGraphContentHasher.swift
@@ -11,16 +11,6 @@ public protocol CacheGraphContentHashing {
     ///     - cacheProfile: Cache profile currently being used
     ///     - cacheOutputType: Output type of cache -> makes a different hash for a different output type
     func contentHashes(
-        for graph: TuistCore.Graph,
-        cacheProfile: TuistGraph.Cache.Profile,
-        cacheOutputType: CacheOutputType
-    ) throws -> [TargetNode: String]
-    /// Hashes graph
-    /// - Parameters:
-    ///     - graph: Graph to hash
-    ///     - cacheProfile: Cache profile currently being used
-    ///     - cacheOutputType: Output type of cache -> makes a different hash for a different output type
-    func contentHashes(
         for graph: ValueGraph,
         cacheProfile: TuistGraph.Cache.Profile,
         cacheOutputType: CacheOutputType
@@ -74,27 +64,6 @@ public final class CacheGraphContentHasher: CacheGraphContentHashing {
     ) -> Bool {
         let isFramework = target.target.product == .framework || target.target.product == .staticFramework
         let noXCTestDependency = !graphTraverser.dependsOnXCTest(path: target.path, name: target.target.name)
-        return isFramework && noXCTestDependency
-    }
-
-    public func contentHashes(
-        for graph: Graph,
-        cacheProfile: TuistGraph.Cache.Profile,
-        cacheOutputType: CacheOutputType
-    ) throws -> [TargetNode: String] {
-        try graphContentHasher.contentHashes(
-            for: graph,
-            filter: filterHashTarget,
-            additionalStrings: [
-                cacheProfileContentHasher.hash(cacheProfile: cacheProfile),
-                cacheOutputType.description,
-            ]
-        )
-    }
-
-    private func filterHashTarget(_ target: TargetNode) -> Bool {
-        let isFramework = target.target.product == .framework || target.target.product == .staticFramework
-        let noXCTestDependency = !target.dependsOnXCTest
         return isFramework && noXCTestDependency
     }
 }