+++ Sources/TuistCore/Graph/Mappers/GraphMapper.swift
@@ -4,33 +4,23 @@ import TuistGraph
 
 /// A protocol that defines an interface to map dependency graphs.
 public protocol GraphMapping {
-    /// Given a graph, it maps it into another graph.
-    /// - Parameter graph: Graph to be mapped.
-    func map(graph: Graph) throws -> (Graph, [SideEffectDescriptor])
-
     /// Given a value graph, it maps it into another value graph.
     /// - Parameter graph: Graph to be mapped.
     func map(graph: ValueGraph) throws -> (ValueGraph, [SideEffectDescriptor])
 }
 
-public extension GraphMapping {
-    func map(graph: ValueGraph) throws -> (ValueGraph, [SideEffectDescriptor]) {
-        return (graph, [])
-    }
-}
-
 /// A mapper that is initialized with a mapping function.
 public final class AnyGraphMapper: GraphMapping {
     /// A function to map the graph.
-    let mapper: (Graph) throws -> (Graph, [SideEffectDescriptor])
+    let mapper: (ValueGraph) throws -> (ValueGraph, [SideEffectDescriptor])
 
     /// Default initializer
     /// - Parameter mapper: Function to map the graph.
-    public init(mapper: @escaping (Graph) throws -> (Graph, [SideEffectDescriptor])) {
+    public init(mapper: @escaping (ValueGraph) throws -> (ValueGraph, [SideEffectDescriptor])) {
         self.mapper = mapper
     }
 
-    public func map(graph: Graph) throws -> (Graph, [SideEffectDescriptor]) {
+    public func map(graph: ValueGraph) throws -> (ValueGraph, [SideEffectDescriptor]) {
         try mapper(graph)
     }
 }
@@ -45,7 +35,7 @@ public final class SequentialGraphMapper: GraphMapping {
         self.mappers = mappers
     }
 
-    public func map(graph: Graph) throws -> (Graph, [SideEffectDescriptor]) {
+    public func map(graph: ValueGraph) throws -> (ValueGraph, [SideEffectDescriptor]) {
         try mappers.reduce((graph, [SideEffectDescriptor]())) { input, mapper in
             let graph = input.0
             var sideEffects = input.1